--- src/main/java/kr/hhplus/be/server/user/domain/PointHistoryRepository.java
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.user.domain;
-
-public interface PointHistoryRepository{
-    void save(PointHistory history);
-}
