+++ src/main/java/kr/hhplus/be/server/user/domain/repository/PointHistoryRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.user.domain.repository;
+
+import kr.hhplus.be.server.user.domain.PointHistory;
+
+public interface PointHistoryRepository{
+    void save(PointHistory history);
+}
