+++ src/main/java/kr/hhplus/be/server/user/infrastructure/PointHistoryRepositoryImpl.java
@@ -1,7 +1,8 @@
 package kr.hhplus.be.server.user.infrastructure;
 
 import kr.hhplus.be.server.user.domain.PointHistory;
-import kr.hhplus.be.server.user.domain.PointHistoryRepository;
+import kr.hhplus.be.server.user.domain.repository.PointHistoryRepository;
+import kr.hhplus.be.server.user.infrastructure.repository.PointHistoryJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
