+++ src/main/java/kr/hhplus/be/server/user/infrastructure/repository/UserJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.user.infrastructure;
+package kr.hhplus.be.server.user.infrastructure.repository;
 
 import kr.hhplus.be.server.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
