+++ src/test/java/kr/hhplus/be/server/testdata/UserTestDataLoader.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.testdata;
 
 import kr.hhplus.be.server.user.domain.User;
-import kr.hhplus.be.server.user.infrastructure.UserJpaRepository;
+import kr.hhplus.be.server.user.infrastructure.repository.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
