+++ src/test/java/kr/hhplus/be/server/user/application/UserServiceIntegrationTest.java
@@ -7,8 +7,8 @@ import kr.hhplus.be.server.testdata.UserTestDataLoader;
 import kr.hhplus.be.server.user.domain.PointHistory;
 import kr.hhplus.be.server.user.domain.User;
 import kr.hhplus.be.server.user.domain.dto.response.PointBalanceResponse;
-import kr.hhplus.be.server.user.infrastructure.PointHistoryJpaRepository;
-import kr.hhplus.be.server.user.infrastructure.UserJpaRepository;
+import kr.hhplus.be.server.user.infrastructure.repository.PointHistoryJpaRepository;
+import kr.hhplus.be.server.user.infrastructure.repository.UserJpaRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
