+++ src/test/java/kr/hhplus/be/server/user/application/UserServiceTest.java
@@ -5,8 +5,8 @@ import kr.hhplus.be.server.common.exception.InsufficientBalanceException;
 import kr.hhplus.be.server.user.domain.PointHistory;
 import kr.hhplus.be.server.user.domain.User;
 import kr.hhplus.be.server.user.domain.dto.response.PointBalanceResponse;
-import kr.hhplus.be.server.user.domain.PointHistoryRepository;
-import kr.hhplus.be.server.user.domain.UserRepository;
+import kr.hhplus.be.server.user.domain.repository.PointHistoryRepository;
+import kr.hhplus.be.server.user.domain.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
